Ensure PWA component updates (#297)
The root `PWA` component dereferenced `title`, `description` and
`canonicalUrl` (also `location` for AMP). Thus mobx would only force
an update when one of those attributes changed. When a project set
all of the attributes to null PWA would never get an update. PWA
updates are necessary for the scroll to top behaviour when navigating.
`scrollResetPending` not being set back to `false` by PWA also made it
so the <Lazy> component would not work upon navigation. As a fix dereference
`app.scrollResetPending`.