Refactor SearchResultProvider state management
This fixes a "flicker" of invalid product images being shown each
time a filter was changed. A local variable was mutated to work around
chaining `updateStore` calls carrying on the old store value. The reference
of the `store` value in the action functions would carry the values of
the initial store. Thus each time a color filter was changed for example
then during the reloading phase the images for the initial products were
shown. Use the functional component equivalent of a `setState` callback,
which is an effect triggering on the relevant store change. Instead of
functions chaining `updateStore` and calling `refresh()` trigger the refresh
in an effect when the store has been updated to `reloading`.
Also remove `appliedFilters`. The combination of `filters` and `reloading`
provides the same functionality.